Add render with notice and render function
