+++ controllers/login.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/foolin/goview/supports/ginview"
 	"github.com/gin-gonic/gin"
 	"github.com/go-playground/validator/v10"
 	errorHandler "github.com/gutakk/go-google-scraper/helpers/error_handler"
+	html "github.com/gutakk/go-google-scraper/helpers/html"
 	session "github.com/gutakk/go-google-scraper/helpers/session"
 	"github.com/gutakk/go-google-scraper/models"
 )
@@ -32,10 +32,7 @@ func (l *LoginController) applyRoutes(engine *gin.RouterGroup) {
 }
 
 func (l *LoginController) displayLogin(c *gin.Context) {
-	ginview.HTML(c, http.StatusOK, loginView, gin.H{
-		"title":   loginTitle,
-		"notices": session.Flashes(c),
-	})
+	html.RenderWithNotice(c, http.StatusOK, loginView, loginTitle, session.Flashes(c), nil)
 }
 
 func (l *LoginController) login(c *gin.Context) {
@@ -64,9 +61,9 @@ func (l *LoginController) login(c *gin.Context) {
 }
 
 func renderLoginWithError(c *gin.Context, status int, err error, form *LoginForm) {
-	ginview.HTML(c, status, loginView, gin.H{
-		"title":  loginTitle,
-		"errors": err.Error(),
-		"email":  form.Email,
-	})
+	data := map[string]interface{}{
+		"email": form.Email,
+	}
+
+	html.RenderWithError(c, status, loginView, loginTitle, err, data)
 }
