+++ templates/login.html
@@ -7,7 +7,7 @@
 {{ include "partials/flashes"}}
 
 <form class="mt-8 space-y-6" action="/login" method="POST">
-  <input name="email" type="email" required placeholder="Email address" value={{.email}}>
+  <input name="email" type="email" required placeholder="Email address" value={{.data.email}}>
   <input name="password" type="password" required placeholder="Password">
 
   <button type="submit">
